requirements.txt
fastapi
sqlalchemy
sqlmodel
uvicorn
python-dotenv
python-jose
passlib
pyodbc